requirements.txt
Django==3.2.1
djangorestframework==3.12.0
